% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{BaseEstimator}
\alias{BaseEstimator}
\title{Title of the R6 Class}
\description{
A brief overview of what this class does.
}
\details{
Add more detailed information about the class.
}
\section{Methods}{

\describe{
    \item{\code{initialize(arg1, arg2)}}{Initializes the class with arguments arg1 and arg2.}
    \item{\code{public_method()}}{Performs a public action or calculation.}
}
}

\section{Private Methods}{

\describe{
    \item{\code{$..private_method()}}{Performs an internal, private action or calculation.}
}
}

\examples{
\dontrun{
    # Create an instance of the class
    obj <- MyClassnew(arg1 = "value", arg2 = 42)

    # Call a public method
    objpublic_method()

    # Access a public field
    print(objfield_name1)
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseEstimator-get_params}{\code{BaseEstimator$get_params()}}
\item \href{#method-BaseEstimator-set_params}{\code{BaseEstimator$set_params()}}
\item \href{#method-BaseEstimator-clone}{\code{BaseEstimator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseEstimator-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-BaseEstimator-get_params}{}}}
\subsection{Method \code{get_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseEstimator$get_params(deep = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Logical. If TRUE, include parameters of nested estimators.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of parameters.
Set parameters of the estimator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseEstimator-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-BaseEstimator-set_params}{}}}
\subsection{Method \code{set_params()}}{
Updates the parameters of the estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseEstimator$set_params(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named parameters to update.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The estimator itself (for chaining).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseEstimator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseEstimator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseEstimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
