% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{DBSCAN}
\alias{DBSCAN}
\title{DBSCAN R6 Class}
\description{
A class for performing DBSCAN clustering.
}
\details{
This class implements the DBSCAN clustering algorithm.
}
\section{Super class}{
\code{\link[DBSCAN615:BaseEstimator]{DBSCAN615::BaseEstimator}} -> \code{DBSCAN}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBSCAN-new}{\code{DBSCAN$new()}}
\item \href{#method-DBSCAN-fit}{\code{DBSCAN$fit()}}
\item \href{#method-DBSCAN-fit_predict}{\code{DBSCAN$fit_predict()}}
\item \href{#method-DBSCAN-plot_clusters}{\code{DBSCAN$plot_clusters()}}
\item \href{#method-DBSCAN-plot_core_samples}{\code{DBSCAN$plot_core_samples()}}
\item \href{#method-DBSCAN-compute_silhouette_score}{\code{DBSCAN$compute_silhouette_score()}}
\item \href{#method-DBSCAN-compute_adjusted_rand_index}{\code{DBSCAN$compute_adjusted_rand_index()}}
\item \href{#method-DBSCAN-compute_noise_ratio}{\code{DBSCAN$compute_noise_ratio()}}
\item \href{#method-DBSCAN-get_labels}{\code{DBSCAN$get_labels()}}
\item \href{#method-DBSCAN-get_n_clusters}{\code{DBSCAN$get_n_clusters()}}
\item \href{#method-DBSCAN-get_core_sample_indices}{\code{DBSCAN$get_core_sample_indices()}}
\item \href{#method-DBSCAN-get_components}{\code{DBSCAN$get_components()}}
\item \href{#method-DBSCAN-get_n_features_in}{\code{DBSCAN$get_n_features_in()}}
\item \href{#method-DBSCAN-get_feature_names_in}{\code{DBSCAN$get_feature_names_in()}}
\item \href{#method-DBSCAN-clone}{\code{DBSCAN$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DBSCAN615" data-topic="BaseEstimator" data-id="get_params"><a href='../../DBSCAN615/html/BaseEstimator.html#method-BaseEstimator-get_params'><code>DBSCAN615::BaseEstimator$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DBSCAN615" data-topic="BaseEstimator" data-id="set_params"><a href='../../DBSCAN615/html/BaseEstimator.html#method-BaseEstimator-set_params'><code>DBSCAN615::BaseEstimator$set_params()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-new}{}}}
\subsection{Method \code{new()}}{
Initializes the DBSCAN class with specified parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$new(
  eps = 0.5,
  min_samples = 5,
  metric = "euclidean",
  metric_params = NULL,
  algorithm = "auto",
  leaf_size = 30,
  p = 2,
  n_jobs = 1,
  max_memory = 1024
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eps}}{float, default=0.5. The maximum distance between two samples for one to be considered in the same neighborhood.}

\item{\code{min_samples}}{int, default=5. The minimum number of samples required to form a cluster.}

\item{\code{metric}}{str or callable, default="euclidean". The distance metric to use.}

\item{\code{metric_params}}{list, default=NULL. Additional arguments for the metric function.}

\item{\code{algorithm}}{str, default="auto". The algorithm used to compute nearest neighbors.}

\item{\code{leaf_size}}{int, default=30. Leaf size passed to BallTree or KDTree.}

\item{\code{p}}{int, default=NULL. Power parameter for Minkowski metric. 1 is Manhattan, 2 is Euclidean, infinity is Chebyshev.}

\item{\code{n_jobs}}{int, default=NULL. The number of parallel jobs to run.}

\item{\code{max_memory}}{int, default=1024. The maximum memory to use for storing the matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-fit"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the DBSCAN clustering model on the input data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$fit(X, y = NULL, sample_weight = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A matrix or data frame. The input data to be clustered.}

\item{\code{y}}{Ignored, present for API consistency.}

\item{\code{sample_weight}}{Numeric vector. Weights for each sample.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The instance itself invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-fit_predict"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-fit_predict}{}}}
\subsection{Method \code{fit_predict()}}{
Fits the DBSCAN clustering model on the input data and returns cluster labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$fit_predict(X, y = NULL, sample_weight = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A matrix or data frame. The input data to be clustered.}

\item{\code{y}}{Ignored, present for API consistency.}

\item{\code{sample_weight}}{Numeric vector. Weights for each sample.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of cluster labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-plot_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-plot_clusters}{}}}
\subsection{Method \code{plot_clusters()}}{
Plots the clusters formed by the DBSCAN algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$plot_clusters(X, title = "DBSCAN Clustering Results")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A matrix or data frame. The input data to be visualized.}

\item{\code{title}}{A string. The title of the plot. Default is "DBSCAN Clustering Results".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object displaying the clusters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-plot_core_samples"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-plot_core_samples}{}}}
\subsection{Method \code{plot_core_samples()}}{
Highlights core samples in the clustering results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$plot_core_samples(X, title = "DBSCAN Core Samples")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A matrix or data frame. The input data to be visualized.}

\item{\code{title}}{A string. The title of the plot. Default is "DBSCAN Core Samples".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object displaying the core samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-compute_silhouette_score"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-compute_silhouette_score}{}}}
\subsection{Method \code{compute_silhouette_score()}}{
Computes the silhouette score for the clustering result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$compute_silhouette_score(X = private$..X, labels = private$..labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A matrix or data frame. The input data used for clustering.}

\item{\code{labels}}{A numeric vector of cluster labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value representing the silhouette score.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-compute_adjusted_rand_index"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-compute_adjusted_rand_index}{}}}
\subsection{Method \code{compute_adjusted_rand_index()}}{
Computes the Adjusted Rand Index (ARI) between true labels and predicted labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$compute_adjusted_rand_index(true_labels, pred_labels = private$..labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_labels}}{A numeric vector. The true cluster labels.}

\item{\code{pred_labels}}{A numeric vector. The predicted cluster labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value representing the ARI score.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-compute_noise_ratio"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-compute_noise_ratio}{}}}
\subsection{Method \code{compute_noise_ratio()}}{
Computes the proportion of noise points in the clustering result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$compute_noise_ratio(labels = private$..labels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{A numeric vector of cluster labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value representing the proportion of noise points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-get_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-get_labels}{}}}
\subsection{Method \code{get_labels()}}{
Retrieves the cluster labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$get_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector of cluster labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-get_n_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-get_n_clusters}{}}}
\subsection{Method \code{get_n_clusters()}}{
Retrieves the number of clusters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$get_n_clusters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer representing the number of clusters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-get_core_sample_indices"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-get_core_sample_indices}{}}}
\subsection{Method \code{get_core_sample_indices()}}{
Retrieves the indices of core samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$get_core_sample_indices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector of core sample indices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-get_components"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-get_components}{}}}
\subsection{Method \code{get_components()}}{
Retrieves the components of core samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$get_components()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix of core sample components.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-get_n_features_in"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-get_n_features_in}{}}}
\subsection{Method \code{get_n_features_in()}}{
Retrieves the number of features in the input data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$get_n_features_in()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer representing the number of features.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-get_feature_names_in"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-get_feature_names_in}{}}}
\subsection{Method \code{get_feature_names_in()}}{
Retrieves the feature names from the input data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$get_feature_names_in()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of feature names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBSCAN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBSCAN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBSCAN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
