% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_adjusted_rand_index.R
\name{metric_adjusted_rand_index}
\alias{metric_adjusted_rand_index}
\title{Adjusted Rand Index Metric}
\usage{
metric_adjusted_rand_index(true_labels, pred_labels)
}
\arguments{
\item{true_labels}{An integer vector of true cluster labels.}

\item{pred_labels}{An integer vector of predicted cluster labels.}
}
\value{
A numeric value representing the ARI.
}
\description{
Computes the Adjusted Rand Index (ARI) for clustering evaluation.
}
