% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_noise_ratio.R
\name{metric_noise_ratio}
\alias{metric_noise_ratio}
\title{Noise ratio}
\usage{
metric_noise_ratio(labels)
}
\arguments{
\item{labels}{An integer vector of cluster labels.}
}
\value{
A numeric value representing the noise ratio.
}
\description{
Computes the noise ratio for clustering evaluation.
}
