% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseEstimator.R
\name{BaseEstimator}
\alias{BaseEstimator}
\title{BaseEstimator R6 Class}
\description{
A base class for implementing estimators in R6.
}
\details{
This class provides utility methods for managing parameters (`get_params` and `set_params`).
It is intended to be inherited by other estimator classes to streamline parameter handling.
}
\section{Methods}{

\describe{
    \item{\code{get_params(deep = TRUE)}}{Retrieves the parameters of the estimator as a named list.}
    \item{\code{set_params(...)}}{Sets the parameters of the estimator, allowing nested parameters.}
}
}

\examples{
\dontrun{
# Create an instance of BaseEstimator (or subclass)
base <- BaseEstimator$new()

# Get parameters
params <- base$get_params()
print(params)

# Set parameters
base$set_params(param1 = "value", param2 = 42)
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseEstimator-get_params}{\code{BaseEstimator$get_params()}}
\item \href{#method-BaseEstimator-set_params}{\code{BaseEstimator$set_params()}}
\item \href{#method-BaseEstimator-clone}{\code{BaseEstimator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseEstimator-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-BaseEstimator-get_params}{}}}
\subsection{Method \code{get_params()}}{
Retrieves the parameters of the estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseEstimator$get_params(deep = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Logical. If TRUE, include parameters of nested estimators.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseEstimator-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-BaseEstimator-set_params}{}}}
\subsection{Method \code{set_params()}}{
Sets the parameters of the estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseEstimator$set_params(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named parameters to update.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The estimator itself (for method chaining).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseEstimator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseEstimator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseEstimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
